1. Turn some ints into bigints.
2. Rename the uniqueness constraint for booking.resource_type.

M    Open-ILS/src/sql/Pg/090.schema.action.sql
M    Open-ILS/src/sql/Pg/200.schema.acq.sql
M    Open-ILS/src/sql/Pg/012.schema.vandelay.sql
M    Open-ILS/src/sql/Pg/095.schema.booking.sql
M    Open-ILS/src/sql/Pg/002.schema.config.sql
M    Open-ILS/src/sql/Pg/070.schema.container.sql
A    Open-ILS/src/sql/Pg/upgrade/0421.schema.embiggen-ints.sql


git-svn-id: svn://svn.open-ils.org/ILS/branches/rel_2_0@18053 dcc99617-32d9-48b4-a31d-7c20da2025e4

+++ Open-ILS/src/sql/Pg/012.schema.vandelay.sql
@@ -77,7 +77,7 @@ ALTER TABLE vandelay.bib_queue ADD PRIMARY KEY (id);
 CREATE TABLE vandelay.queued_bib_record (
 	queue		INT		NOT NULL REFERENCES vandelay.bib_queue (id) ON DELETE CASCADE DEFERRABLE INITIALLY DEFERRED,
 	bib_source	INT		REFERENCES config.bib_source (id) DEFERRABLE INITIALLY DEFERRED,
-	imported_as	INT		REFERENCES biblio.record_entry (id) DEFERRABLE INITIALLY DEFERRED
+	imported_as	BIGINT	REFERENCES biblio.record_entry (id) DEFERRABLE INITIALLY DEFERRED
 ) INHERITS (vandelay.queued_record);
 ALTER TABLE vandelay.queued_bib_record ADD PRIMARY KEY (id);
 CREATE INDEX queued_bib_record_queue_idx ON vandelay.queued_bib_record (queue);
